Huge commit! - EMS Visplanes! they are designed to be grouped as many as possible per frame, and to only take up one frame at a time. this is made possible by... - Page Locking! Now pages can be locked in memory with a certain variable. This is made possible with internal reference counting on the locked pages.
EMS visplanes work with just 4 EMS pages, but it makes things significantly slower. But the idea is we can take that 85k of conventional memory back and use it on something more useful (?). For now, i will leave EMS visplanes on.

I expect to find some big positive performance gains in a few spots of the codebase with the new page locking feature to prevent page thrashing in certain game loops. Also, with 8 EMS pages available some more aggressive locking may help. There are some #defines currently in the codebase based on this... i don't know how much I wish to litter the codebase with that though.